% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rule_building.R
\name{add_negative_terms}
\alias{add_negative_terms}
\title{pos/neg' = glue('{tot_pos}/{tot_neg}'),}
\usage{
add_negative_terms(rules, target_vec, target_data)
}
\description{
pos/neg' = glue('{tot_pos}/{tot_neg}'),
}
