% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Record_management.R
\name{import_data}
\alias{import_data}
\title{Read a data file/object}
\usage{
import_data(input, ...)
}
\arguments{
\item{input}{A file path to an Excel/CSV file or a data frame object}

\item{...}{Additional arguments to pass to
\code{link[readxl:read_excel]{readxl::read_excel}} or
\code{link[readr:read_csv]{readr::read_csv}}, depending on the file type.}
}
\value{
A data frame object.
}
\description{
Read a data file/object
}
\examples{

\dontrun{
# These all work:

data <- import_data('data.csv')

data <- import_data('data.xlsx', sheet = 2)

data <- import_data(data)
}
}
